fix(documents): evitar 500 en la lista por desborde de bigint y mostrar el error real
El ORDER BY de GET /api/documents casteaba el transmittal numerico a ::bigint.
Si algun transmittal tenia mas de 19 digitos, el cast desbordaba y Postgres
lanzaba error: la lista entera fallaba con 500 aunque los reportes (que no
hacen ese cast) si cargaran. Se cambia a ::numeric (precision arbitraria, sin
desborde) manteniendo el mismo orden numerico.

Ademas, getDocuments en el frontend ocultaba el error del servidor tras un
texto generico; ahora propaga el mensaje real para poder diagnosticar fallos
de la consulta desde la UI.

+++ backend/routes/documents.js
@@ -71,7 +71,7 @@ router.get('/', async (req, res) => {
       LEFT JOIN contracts c ON c.code = documents.n_contrato AND c.organization_id = documents.organization_id`
       + (conds.length ? ' WHERE ' + conds.join(' AND ') : '')
       + ` ORDER BY fecha ASC NULLS LAST,
-          NULLIF(regexp_replace(transmittal, '[^0-9]', '', 'g'), '')::bigint ASC NULLS LAST,
+          NULLIF(regexp_replace(transmittal, '[^0-9]', '', 'g'), '')::numeric ASC NULLS LAST,
           transmittal ASC NULLS LAST, id ASC`;
     const { rows } = await pool.query(query, params);
     res.json(rows);
